move more handlers around
